requirements.txt
Biopython
pytest